fix: address PR #30 review feedback
- Add int32_t to JitResult variant and reorder by ascending width
- Always resolve entry function return type for correct ABI calling convention
- Return int32_t directly from execute() instead of casting to int64_t

+++ compiler/src/jit_session.cpp
@@ -171,7 +171,7 @@ std::optional<JitResult> JITSession::execute(const std::string& entryName,
   }
   if (resultType == "i32" || resultType == "u32") {
     auto entryFn = sym->toPtr<int32_t (*)()>();
-    return static_cast<int64_t>(entryFn());
+    return entryFn();
   }
   auto entryFn = sym->toPtr<int64_t (*)()>();
   return entryFn();
